Bugfix: colormap for heatmap is now correct across tiles.
New feature: the data extraction from a nanocube is now parameterizable. There are two functions. First is process_values, passed to the Nanocube construction, which converts the raw vector returned from a nanocube to whatever will be used (regression coefficients, etc). The second function is 'value_function', which takes the result of process_values and produces a number to be shown by the bar charts and time series plots.

TODO: value_function needs to be split into many different functions, one per visualization, in order for us to be able to show different aspects of the data in different plots.